move existing auth flows to auth controller
