+++ client/src/api/useRefreshToken.ts
@@ -2,13 +2,13 @@ import { useAuth } from '../context/AuthProvider';
 import { refreshHttpInstance } from './httpService';
 
 const useRefreshToken = () => {
-    const { setAuthToken } = useAuth();
+    const { setAccessToken } = useAuth();
 
     const refresh = async () => {
         const response = await refreshHttpInstance.post('/refresh-token', {
             withCredentials: true
         });
-        setAuthToken(response.data.accessToken);
+        setAccessToken(response.data.accessToken);
         return response.data.accessToken;
     }
     return refresh;
